requirements.txt
discord-timers==0.1.0
discord.py==2.0.0
mutagen==1.45.1
nest-asyncio==1.5.5
python-dotenv==0.20.0
PyNaCl == 1.5.0
mcstatus == 9.3.1
mcrcon == 0.7.0